refactor: extract stake/unstake mode switcher from AppLayout
