+++ src/features/stake/Stake.tsx
@@ -1,5 +1,6 @@
 import { useStaking } from 'src/features/stake/useStaking';
 import { SwapForm } from 'src/features/swap/SwapForm';
+import { Switcher } from 'src/features/swap/Switcher';
 import { toCeloWei } from 'src/formatters/amount';
 import { useAccountContext } from 'src/providers/AccountProvider';
 import toast from 'src/services/toast';
@@ -17,6 +18,7 @@ export const Stake = () => {
 
   return (
     <div className="flex flex-col md:w-96 mx-auto w-full px-4 mb-14">
+      <Switcher />
       <SwapForm
         estimateReceiveValue={estimateDepositValue}
         estimateGasFee={estimateStakingFee}
