+++ src/features/unstake/Unstake.tsx
@@ -1,4 +1,5 @@
 import { SwapForm } from 'src/features/swap/SwapForm';
+import { Switcher } from 'src/features/swap/Switcher';
 import { useUnstaking } from 'src/features/unstake/useUnstaking';
 import { toStCeloWei } from 'src/formatters/amount';
 import { useAccountContext } from 'src/providers/AccountProvider';
@@ -19,6 +20,7 @@ export const Unstake = () => {
 
   return (
     <div className="flex flex-col md:w-96 mx-auto w-full px-4 mb-14">
+      <Switcher />
       <SwapForm
         estimateReceiveValue={estimateWithdrawalValue}
         estimateGasFee={estimateUnstakingFee}
