+++ src/layout/AppLayout.tsx
@@ -30,7 +30,7 @@ export const AppLayout = ({ pathName, children }: PropsWithChildren<Props>) => {
         <title>{`Liquid Staking | ${getHeadTitle(pathName)}`}</title>
       </Head>
       <div className="flex flex-col h-full min-h-screen w-full min-w-screen">
-        {isConnected && <Header pathName={pathName} />}
+        {isConnected && <Header />}
         <main className="w-full flex-1 flex">{children}</main>
         <Footer />
       </div>
